Add files via upload
考古化学源代码文件
Volume, as one of the important characteristics of cultural relics, can reflect the shape of the relics and even their geographical, technological and cultural sequence. In the field of archaeology, calculating the volume of cultural relics is a crucial task. The traditional method of volume measurement relies on manual measurement and traditional mapping tools, which is time-consuming and labour-intensive, and it is difficult to provide sufficiently accurate and comprehensive data for irregular artifacts. To achieve more accurate and efficient measurement of artifacts, this paper proposes a multi-view 3D reconstruction method based on COLMAP software. By simply taking a set of overlapping images of artifacts and known volume references, using COLMAP software to generate a point cloud model, and using the known volume of the reference in the image system to establish the proportionality between the volume of the point cloud model and the volume of the real artifacts, and then calculate the volume of the artifacts. By optimizing the scene model and exploring the parameters of the COLMAP software, this method achieves a difference of less than 5% between accuracy and the traditional method in about ten minutes of response time, providing a convenient and fast method for measuring the volume of artifacts in archaeology.